# 스킬 쿨타임, MP감소, 데미지 적용
- 스킬 쿨타임 적용 완료
- 스킬마다 감소되는 MP양과, 입혀지는 데미지 개별로 적용 완료
- 스킬 실행중일때 다른 스킬 실행안되도록 예외처리 추가해야함

+++ Assets/KYJ/Removeeeeeee/EnemyControl.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyControl : MonoBehaviour
 {
-    public int curHp = 100;
+    public float curHp = 100.0f;
 
     private void Update()
     {
@@ -12,7 +12,7 @@ public class EnemyControl : MonoBehaviour
             gameObject.SetActive(false);
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         curHp -= damage;
         print(this.name+" Damaged! - " +curHp);
